Move AudioBuffer definition to the Audio.d.ts and remove the previous import.
